refactor: move challenge DTOs to a dedicated namespace
