+++ TCSA.V2026/Data/DTOs/Challenges/MarkChallengeCompletedRequest.cs
@@ -1,3 +1,3 @@
-namespace TCSA.V2026.Data.DTOs;
+namespace TCSA.V2026.Data.DTOs.Challenges;
 
 public record MarkChallengeCompletedRequest(int ChallengeId, string UserId);
