+++ TCSA.V2026/Data/DTOs/Challenges/SyncChallengeRequest.cs
@@ -1,3 +1,3 @@
-namespace TCSA.V2026.Data.DTOs;
+namespace TCSA.V2026.Data.DTOs.Challenges;
 
 public record SyncChallengeRequest(UserPlatformCredentials PlatformCredentials, int ChallengeId, string ExternalId, string UserId);
